Added Corrected SpriteSheet Of The Meerkat 1. Added The New Spritesheet Of The Meerkat With His New Collision Box. 2. New Organization System Of The Game Folders Starting With AI

+++ ModelingProject1/SourceCode/PushButton.cpp
@@ -16,7 +16,7 @@ RPRGUI::PushButton::~PushButton(void)
 
 int RPRGUI::PushButton::eventClicked(bool* isRunning)
 {
-  if ( idChangeState == GameStates::STATE_EXIT )
+  if ( idChangeState == (int)GameStates::STATE_EXIT )
   {
     *isRunning = false;
   }
